Rename eBPF collector (#8822)
We renamed eBPF collector for a more meaningful name.
