+++ .gitignore
@@ -66,8 +66,8 @@ slabinfo.plugin
 cgroup-network
 !cgroup-network/
 
-ebpf_process.plugin
-!ebpf_process.plugin/
+ebpf.plugin
+!ebpf.plugin/
 
 # protoc generated files
 *.pb.cc
