+++ collectors/ebpf.plugin/Makefile.am
@@ -8,5 +8,6 @@ dist_noinst_DATA = \
     $(NULL)
 
 dist_libconfig_DATA = \
-    ebpf_process.conf \
+    ebpf.conf \
     $(NULL)
+
