+++ collectors/ebpf.plugin/ebpf.conf
